refactor(ts): service module users routes
